Push app-specific logic into app package
