+++ packages/app/src/cli/services/dev.ts
@@ -15,9 +15,10 @@ import {UIExtension} from '../models/app/extensions.js'
 import {fetchProductVariant} from '../utilities/extensions/fetch-product-variant.js'
 import {load} from '../models/app/loader.js'
 import {getAppIdentifiers} from '../models/app/identifiers.js'
+import {getAnalyticsTunnelType} from '../utilities/analytics.js'
 import {output, system, session, abort, string, environment} from '@shopify/cli-kit'
 import {Config} from '@oclif/core'
-import {getAnalyticsTunnelType, reportEvent} from '@shopify/cli-kit/node/analytics'
+import {reportEvent} from '@shopify/cli-kit/node/analytics'
 import {execCLI2} from '@shopify/cli-kit/node/ruby'
 import {renderConcurrent} from '@shopify/cli-kit/node/ui'
 import {getAvailableTCPPort} from '@shopify/cli-kit/node/tcp'
