+++ packages/cli-kit/src/public/node/analytics.ts
@@ -4,9 +4,9 @@ import {content, debug, token} from '../../output.js'
 import constants from '../../constants.js'
 import * as metadata from '../../metadata.js'
 import {publishEvent, MONORAIL_COMMAND_TOPIC} from '../../monorail.js'
-import {fanoutHooks, getListOfTunnelPlugins} from '../../plugins.js'
+import {fanoutHooks} from '../../plugins.js'
 import {getEnvironmentData, getSensitiveEnvironmentData} from '../../private/node/analytics.js'
-import {Config, Interfaces} from '@oclif/core'
+import {Interfaces} from '@oclif/core'
 
 interface ReportEventOptions {
   config: Interfaces.Config
@@ -42,28 +42,6 @@ export async function reportEvent(options: ReportEventOptions): Promise<void> {
     }
     debug(message)
   }
-}
-
-/**
- * Return the name of the tunnel provider used to send analytics. Returns 'localhost' or provider name if any of those
- * strings are included in the {@link tunnelUrl} param. Returns 'custom' otherwise
- *
- * @param options - Oclif configuration. Needed to call the hook for retrieving the list of tunner providers
- * @param tunnelUrl - Tunnel url. Used as pattern to match provider name
- * @returns 'localhost' or provider name if any of those strings are included in
- *  the tunnelUrl or 'custom' otherwise
- */
-export async function getAnalyticsTunnelType(options: Config, tunnelUrl: string): Promise<string | undefined> {
-  if (!tunnelUrl) {
-    return
-  }
-
-  if (tunnelUrl.includes('localhost')) {
-    return 'localhost'
-  }
-
-  const provider = (await getListOfTunnelPlugins(options)).plugins.find((plugin) => tunnelUrl?.includes(plugin))
-  return provider ?? 'custom'
 }
 
 async function buildPayload({config, errorMessage}: ReportEventOptions) {
